# Optimize some inner cases for tokenFromKeyword