Some nested cases which contain a lot of comparisons (>= 5) with long keyword lengths and common prefixes could be optimized more. WDYT?

**benchmark**
Parse AS.
main branch:
```
Parse      :   785.185 ms  n=171
Parse      :   768.248 ms  n=171
Parse      :   801.483 ms  n=171
Parse      :   776.191 ms  n=171
```

This PR:
```
Parse      :   727.563 ms  n=171
Parse      :   759.355 ms  n=171
Parse      :   760.821 ms  n=171
Parse      :   769.710 ms  n=171
```

- [x] I've read the contributing guidelines
- [x] I've added my name and email to the NOTICE file